shellScript to run all
This shell script will run all the integrated parts of the flow simulation and permeability prediction. Once you run the XCT code the info regarding each block will be created and sent to individual folders. The user needs to locate the present working directory for the rest of the simulation in the Permeability folder.